fix(app-shell): error link to support token retry
